Moved granule view stuff to a subdirectory.
