+++ client/src/detail/GranuleTab/list.css
@@ -1,0 +1,191 @@
+@import '../../../style/style.css';
+/*
+@define-extend underscored {
+  border-bottom: 1px solid $color-gray;
+  padding-bottom: 1em;
+  margin-bottom: 1em;
+}*/
+
+.mainWindow {
+  letter-spacing: 0 !important;
+  background-color: #000;
+  margin: 2em auto;
+  max-width: 100em;
+  height: calc(100vh - 5em - $header-height - $footer-height*2);
+}
+
+/*.navLink {
+  color: #00a6d2;
+  font-size: 1.15em;
+  text-decoration: underline;
+  cursor: pointer;
+}*/
+
+.mapContainer {
+  height: 100%;
+  width: 100%;
+}
+
+.granule,
+.map {
+  height: 100%
+}
+
+.title {
+  @extend underscored;
+  font-size: 1.5em;
+  padding-top: 1em;
+}
+
+/*.description {
+  @extend underscored;
+  display: inline-block;
+  font-size: 1em;
+  vertical-align: top;
+}*/
+
+.granuleInfo {
+  padding: 0 1em;
+  height: 100%;
+  overflow-y: auto;
+}
+
+.table {
+  width: 100%;
+  tbody {
+    tr:hover {
+      background-color: #144CBD;
+    }
+  }
+}
+
+@define-mixin badge $size: 1em {
+  border-radius: 50%;
+  width: $size !important;
+  height: $size !important;
+  padding: calc($size / 4) !important;
+  margin: calc($size / 4) !important;
+  font: $size Arial, sans-serif;
+  color: white;
+  text-align: center;
+  text-decoration: none;
+}
+
+.legend {
+  justify-content: flex-end;
+}
+
+.legend {
+  display: flex;
+  flex-flow: row wrap;
+  padding-bottom: 1em;
+  h1, h2, h3, h4, h5 {
+    margin: 0.6em 0 !important;
+  }
+  .legendItem {
+    display: inline-flex;
+    margin-left: 1em;
+    .badge {
+      @mixin badge 1.2em
+    }
+    .label {
+      font: 1.2em Arial, sans-serif;
+      margin: 0.6em 0 !important;
+    }
+  }
+}
+
+/*.granuleList {
+  list-style-type: none;
+  padding: 0;
+  margin: 0;
+}*/
+
+.badgeCell {
+  width: 25%;
+  min-width: 4em;
+  .badgeLayout {
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: flex-end;
+    .badge {
+      @mixin badge
+    }
+  }
+}
+
+tr.pageButton {
+  text-align: center;
+  border-top: 1px solid #cbcbcb;
+  button {
+    color: white;
+    background-color: $color-primary;
+    text-align: center;
+    font-size: 1.25em;
+  }
+}
+
+/* Section 508 style */
+/*.descriptionContainer {
+  padding: 1em;
+  background: $color-primary-darker;
+}*/
+
+/*.descriptionParagraph {
+  line-height: 1.4em;
+}
+
+.listContainer {
+  box-shadow: 1.3em 0.7em 0.5em 0.2em rgba(0,0,0,.7);
+  overflow-x: none;
+}
+
+.granuleList508 {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+  a {
+    color: white;
+  }
+}*/
+
+/*.listItem {
+  padding: .5em 1em 1em;
+  > div {
+    margin: .5em 0;
+  }
+  &:nth-child(odd)	{
+    background-color: $color-primary-darkest;
+  }
+  &:nth-child(even) {
+    background-color: $color-primary-darker;
+  }
+}
+
+.linkRow {
+  display: inline-block;
+  margin: .5em .5em .5em 0;
+}*/
+
+.links {
+  display: inline-block;
+  padding: .3em;
+  margin: .2em;
+  border-radius: .1em .4em;
+  color: white;
+  background-color: $color-green;
+  text-decoration: underline;
+  cursor: pointer;
+}
+/*.showMore {
+  width: 100%;
+  text-align: center;
+  button {
+    padding: .3em;
+    margin: .2em;
+    border: none;
+    border-radius: .1em .4em;
+    color: white;
+    background-color: $color-primary;
+  }
+}*/
